+++ src/store/partials/api.js
@@ -1,17 +1,20 @@
 
 export default {
   state: {
-    apiURL: {
+    url: {
       JOIN: '/user/signup/',
+      VALIDATE_EMAIL: '/user/validate_email/',
       LOGIN: '/user/login/',
       LOGOUT: '/user/logout/',
-      USER_LIST: '/user/',
-      GROUP_LIST: '/group/all/',
+      USER_PROFILE: '/user/profile/',
       GROUP_REGISTER: '/group/register/',
+      HOBBY_LIST: '/group/hobby/',
 
     },
   },
   getters: {
-
+    url(state) {
+      return state.url;
+    },
   },
 };
