+++ src/store/partials/region.js
@@ -1,0 +1,24 @@
+import axios from 'axios';
+
+let http = axios;
+
+export default {
+  state: {
+    regionList: [],
+  },
+  getters: {
+    regionList(state) {
+      return state.regionList;
+    },
+  },
+  mutations: {
+    setRegionList() {
+      
+    },
+  },
+  actions: {
+    getRegionList({commit}) {
+
+    },
+  }
+};
